Set value of non-checkout events to 0, use price custom property instead.
Value means the value of the event to track, for product page views and add
to cart events this value is 0, only for checkout events value should be
the actual value of the order. Price custom data has been introduced to
keep track of products viewed and added to cart prices.